server/requirements.txt
openenv-core>=0.2.2
huggingface_hub>=0.24.0
fastapi>=0.115.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0
networkx>=3.0
requests>=2.31.0
httpx>=0.27.0
websockets>=12.0
openai>=1.0.0
